+++ Funcky/Extensions/EnumerableExtensions.cs
@@ -67,7 +67,7 @@ namespace Funcky.Extensions
         /// The IEnumerable version of foreach. You can apply an action to each element. This is only useful when you have side effects.
         /// </summary>
         /// <typeparam name="T">the inner type of the enumerable.</typeparam>
-        public static void Each<T>(this IEnumerable<T> elements, Action<T> action)
+        public static void ForEach<T>(this IEnumerable<T> elements, Action<T> action)
         {
             foreach (var element in elements)
             {