+++ changelog.md
@@ -14,4 +14,7 @@
 * Add `FirstOrNone`, `LastOrNone` and `SingleOrNone` extension functions
 
 ## Next major version
+* Use `Func<T, bool>` instead of `Predicate<T>` in predicate composition functions (`Functional.All`, `Functional.Any`, `Functional.Not`),
+  because most APIs in `System` use `Func`.
+* `Functional.Any` now returns `false` when the given list of predicates is empty.
 * The `Each` extension method on `IEnumerable<T>` has been renamed to `ForEach`.