+++ changelog.md
@@ -15,3 +15,4 @@
 
 ## Unreleased
 * Added overload for AndThen which flattens the Option
+* Add `Where` method to `Option<T>`, which allows filtering the `Option` by a predicate.