+++ changelog.md
@@ -14,6 +14,7 @@
 * Add `FirstOrNone`, `LastOrNone` and `SingleOrNone` extension functions
 
 ## Next major version
+* Rename `Option.From` -> `Option.FromNullable` and remove overload that takes non-nullable value types.
 * Use `Func<T, bool>` instead of `Predicate<T>` in predicate composition functions (`Functional.All`, `Functional.Any`, `Functional.Not`),
   because most APIs in `System` use `Func`.
 * `Functional.Any` now returns `false` when the given list of predicates is empty.