+++ changelog.md
@@ -18,3 +18,4 @@
 * Use `Func<T, bool>` instead of `Predicate<T>` in predicate composition functions (`Functional.All`, `Functional.Any`, `Functional.Not`),
   because most APIs in `System` use `Func`.
 * `Functional.Any` now returns `false` when the given list of predicates is empty.
+* The `Each` extension method on `IEnumerable<T>` has been renamed to `ForEach`.