+++ changelog.md
@@ -16,3 +16,5 @@
 ## Unreleased
 * Added overload for AndThen which flattens the Option
 * Add `Where` method to `Option<T>`, which allows filtering the `Option` by a predicate.
+* Add overload for `Option<T>.SelectMany` that takes only a selector.
+